Move UncaughtExceptionSource out of worker.c++
